requirements.txt
yacs==0.1.8
numpy==1.21.2
torchvision
tqdm==4.62.3
opencv_contrib_python==4.5.3.56
opencv-contrib-python-headless
imageio==2.9.0
torch
mmcv_full
matplotlib==3.4.3
mmcv
Pillow==8.4.0
scikit_learn==1.0.1

